add kp-dump-xml
Closes #35

Add a utility library for dumping the internal KDBX XML representation
so that parsers can be more easily written.

In order to not include usually-unnecessary dependencies, a new feature
flag 'utilities' was introduced that needs to be set in order to run the
new tool, e.g.

cargo run --release --features "utilities" --bin kp-dump-xml -- -h