Refactoring of group members
Related to #16 #24

In order to support multiple Groups or Entries with the same name at the
same hierarchy level, make several changes to the internal
representation of databases.

Groups now contain children, a Vec<Node> . Node is an enum with Group or
Entry variants containing the child data. What was called Node before is
now called NodeRef (a reference to within a database with a lifetime).
There now also is a NodeRefMut for mutable references into the database.