refactor(xml)!: re-write XML parsing
In preparation for adding additional fields from the XML document for
reading and writing, rewrite the XML parsing code from an explicitly
managed stack to use a call stack together with the new FromXml trait.

BREAKING CHANGE: There are some differences in the `times` fields of
`Entry` and `Group` types, and some of the internal errors might have
changed structure.

BREAKING CHANGE: Currently, all unrecognized fields in the inner XML
database will lead to an error. This is to aid development, but should
probably be softened to a warning for the sake of background
compatibility.