refactor(xml)!: split into modules (#97)
In preparation for adding additional XML fields to the `Entry` and `Group`,
`Meta`, and other types, clean up the parsing and dumping code.

This commit moves the code for XML parsing and dumping into separate 
modules, and rewrites the code to be based around new `FromXml` and 
`DumpXml` traits to reduce repetitive code.

The new XML parsing code migrates from an explicitly managed stack to 
use a call stack. This means that the parsers have to make sure that they 
consume all of the events pertaining to a tag and may only return after a 
tag that was opened in their function has also been closed again. If there 
is a tag that the parser does not recognize,  it should pass execution into 
a `IgnoreSubfield` sub-parser that consumes events until we have 
returned from that tag.

BREAKING CHANGE: There are some differences in the `times` fields of
`Entry` and `Group` types, and some of the internal errors might have
changed structure.