fix: set all default times when creating a node (#154)
Other KDBX parsers (like the one in KeePassXC)
store all the default times as fields of a class,
instead of items in a map. The result is that
default times that are not populated will be
given a default value at parsing time. This can
result in drift between an entry and its history items,
as described in https://github.com/sseemayer/keepass-rs/issues/153.
Populating all the default times is thus safer
for interoperability between KDBX4 parsers.