fix: ensure unique modification timestamps for history entries
Previously, the last modification timestamp would
be updated only after committing a new entry history
item. This resulted in history entries with the
same last modification timestamp if an entry was
modified from two different places. While this is not
a problem now, it will be problematic when we
implement a merge/sync functionality.