feat: add support for challenge-response keys
This commit adds the basic structure for challenge-response keys,
as well as a `LocalChallenge` version of the
`ChallengeResponseKey`, which can be used to recover a database using
the HMAC-SHA1 secret directly.

Currently, the function to perform the challenge is synchronous, but
I think in the future we might also want to add an async version, so
that the feature can be used without blocking the clients' UIs.