feat: implement zeroize for DatabaseKey
I also added a utility function to make sure that the password
never leaves the DatabaseKey when using a password prompt. We might
consider exposing this feature from the crate at some point, but for
now the `utilities` feature has to be enabled.

Addresses https://github.com/sseemayer/keepass-rs/issues/171