feat: add DatabaseKey::is_empty fn
This also fixes a bug with the `with_password_from_prompt` function that
prevented users from entering an empty password.

BREAKING CHANGE: an empty password is no longer converted to a database
key with no password when using the command-line utilities. The
`--no-password` option has to be used to open a database without a
password.