fix(test): feature-gate roundtrip tests (#249)
The roundtrip tests can only succeed if the database to be parsed in the
test has been successfully generated at least one time, so tests will
fail when the project was freshly cloned and the tests are run with the
database missing.

Put the whole roundtrip test module behind a feature gate, so that we
test writing and reading at the same time.

Co-authored-by: louib <code@louib.net>